requirements/local.txt
-r base.txt

django-pdb
django-debug-toolbar
